Dad would be proud. <3

+++ projects/stooper/scripts/migrate.tsx
@@ -1,9 +1,11 @@
 import { resolve } from "@std/path";
+import { existsSync } from "@std/fs/exists";
 
 import { $ } from "@brainbow/ethos/dev/shell";
 import * as sh from "@brainbow/ethos/dev/shell";
+import type { Args } from "@brainbow/ethos/dev/shell";
 
-const args = sh.parse(Deno.args);
+const args = sh.parse<Args>(Deno.args);
 
 args.workdir ??= resolve(import.meta.dirname!, "..");
 args.reset ??= false;
@@ -21,7 +23,7 @@ if (await sh.which("sqlite3") == undefined) {
     throw new Error(`Couldn't find SQLite installed. Have you bootstrapped your workspace yet?`);
 }
 
-if (args.reset == true) {
+if (args.reset == true && existsSync(catalogDbPath)) {
     console.log(`Got reset; nuking database '${catalogDbPath}`);
     await Deno.remove(catalogDbPath);
 }
